Make it possible to read bool in log

+++ classes/GUI/class.xvmpContentGUI.php
@@ -33,7 +33,7 @@ class xvmpContentGUI extends xvmpGUI
                     $mid = $_GET['mid'];
                 }
                 if (!$mid || !xvmpSelectedMedia::isSelected($mid, $this->getObjId())) {
-                    $is_selected = xvmpSelectedMedia::isSelected($mid, $this->getObjId());
+                    $is_selected = (int) xvmpSelectedMedia::isSelected($mid, $this->getObjId());
                     xvmpCurlLog::getInstance()->write(sprintf('Access denied: no mid found for this obj_id %s (is selected: %s), with following command %s', $this->getObjId(), $is_selected, self::CMD_RENDER_LIST_ITEM));
                 }
                 break;
